+++ supertokens_python/recipe/thirdpartyemailpassword/emaildelivery/services/backward_compatibility/__init__.py
@@ -27,11 +27,11 @@ from supertokens_python.recipe.emailpassword.utils import \
     InputResetPasswordUsingTokenFeature
 from supertokens_python.recipe.emailverification.emaildelivery.services.backward_compatibility import \
     BackwardCompatibilityService as EVBackwardCompatibilityService
-from supertokens_python.recipe.emailverification.types import User as EVUser, VerificationEmailTemplateVars
+from supertokens_python.recipe.emailverification.types import User as EVUser
 from supertokens_python.recipe.thirdpartyemailpassword.interfaces import \
     RecipeInterface
-from supertokens_python.recipe.thirdpartyemailpassword.types import \
-    ThirdPartyEmailPasswordEmailTemplateVars
+from supertokens_python.recipe.thirdpartyemailpassword.types import (
+    EmailTemplateVars, VerificationEmailTemplateVars)
 from supertokens_python.supertokens import AppInfo
 
 if TYPE_CHECKING:
@@ -39,7 +39,7 @@ if TYPE_CHECKING:
         InputEmailVerificationConfig
 
 
-class BackwardCompatibilityService(EmailDeliveryInterface[ThirdPartyEmailPasswordEmailTemplateVars]):
+class BackwardCompatibilityService(EmailDeliveryInterface[EmailTemplateVars]):
     ep_backward_compatiblity_service: EPBackwardCompatibilityService
     ev_backward_compatiblity_service: EVBackwardCompatibilityService
 
@@ -75,7 +75,7 @@ class BackwardCompatibilityService(EmailDeliveryInterface[ThirdPartyEmailPasswor
             ep_email_verification_feature,
         )
 
-    async def send_email(self, template_vars: ThirdPartyEmailPasswordEmailTemplateVars, user_context: Dict[str, Any]) -> None:
+    async def send_email(self, template_vars: EmailTemplateVars, user_context: Dict[str, Any]) -> None:
         if isinstance(template_vars, VerificationEmailTemplateVars):
             await self.ev_backward_compatiblity_service.send_email(template_vars, user_context)