+++ supertokens_python/recipe/thirdpartypasswordless/__init__.py
@@ -27,8 +27,8 @@ from . import utils
 from .emaildelivery import services as emaildelivery_services
 from .recipe import ThirdPartyPasswordlessRecipe
 from .smsdelivery import services as smsdelivery_services
-from .types import (ThirdPartyPasswordlessEmailTemplateVars,
-                    ThirdPartyPasswordlessSMSTemplateVars)
+from .types import (EmailTemplateVars,
+                    SMSTemplateVars)
 
 InputEmailVerificationConfig = utils.InputEmailVerificationConfig
 InputOverrideConfig = utils.InputOverrideConfig
@@ -60,8 +60,10 @@ def init(contact_config: ContactConfig,
              PhoneOrEmailInput, Dict[str, Any]], Awaitable[str]], None] = None,
          get_custom_user_input_code: Union[Callable[[Dict[str, Any]], Awaitable[str]], None] = None,
          email_verification_feature: Union[InputEmailVerificationConfig, None] = None,
-         email_delivery: Union[EmailDeliveryConfig[ThirdPartyPasswordlessEmailTemplateVars], None] = None,
-         sms_delivery: Union[SMSDeliveryConfig[ThirdPartyPasswordlessSMSTemplateVars], None] = None,
+         email_delivery: Union[EmailDeliveryConfig[EmailTemplateVars], None] = None,
+         sms_delivery: Union[SMSDeliveryConfig[SMSTemplateVars], None] = None,
          override: Union[InputOverrideConfig, None] = None,
          providers: Union[List[Provider], None] = None) -> Callable[[AppInfo], RecipeModule]:
-    return ThirdPartyPasswordlessRecipe.init(contact_config, flow_type, get_link_domain_and_path, get_custom_user_input_code, email_verification_feature, email_delivery, sms_delivery, override, providers)
+    return ThirdPartyPasswordlessRecipe.init(contact_config, flow_type, get_link_domain_and_path,
+                                             get_custom_user_input_code, email_verification_feature,
+                                             email_delivery, sms_delivery, override, providers)