+++ supertokens_python/recipe/thirdpartypasswordless/smsdelivery/services/twilio/__init__.py
@@ -21,15 +21,15 @@ from supertokens_python.ingredients.smsdelivery.types import \
 from supertokens_python.recipe.passwordless.smsdelivery.services.twilio import \
     TwilioService as PlessTwilioService
 
-from ....types import ThirdPartyPasswordlessSMSTemplateVars
+from ....types import SMSTemplateVars
 
 
-class TwilioService(SMSDeliveryInterface[ThirdPartyPasswordlessSMSTemplateVars]):
+class TwilioService(SMSDeliveryInterface[SMSTemplateVars]):
     pless_twilio_service: PlessTwilioService
 
     def __init__(self, twilio_settings: TwilioSettings,
-                 override: Union[Callable[[TwilioServiceInterface[ThirdPartyPasswordlessSMSTemplateVars]], TwilioServiceInterface[ThirdPartyPasswordlessSMSTemplateVars]], None] = None) -> None:
+                 override: Union[Callable[[TwilioServiceInterface[SMSTemplateVars]], TwilioServiceInterface[SMSTemplateVars]], None] = None) -> None:
         self.pless_twilio_service = PlessTwilioService(twilio_settings, override)
 
-    async def send_sms(self, template_vars: ThirdPartyPasswordlessSMSTemplateVars, user_context: Dict[str, Any]) -> None:
+    async def send_sms(self, template_vars: SMSTemplateVars, user_context: Dict[str, Any]) -> None:
         await self.pless_twilio_service.send_sms(template_vars, user_context)